+++ CHANGES.md
@@ -18,6 +18,8 @@
 - `info()`, `debug()`, and `warning()` now take a variable number of arguments
   to print
 - Add support for setting the entry point for native binaries (C, C++, etc)
+- Regenerating build files with the Ninja backend now uses the `console` pool,
+  allowing realtime output and colored text
 
 ### Breaking changes
 - Drop support for Python 2